or1200_genpc: fix ipcu_cycstb_o generation
In some circumstances the CPU is still waiting for the lsu to finish
while in a pre branch state. However, ipcu_cycstb_o is set and the cycle
starts with the wrong address on the iwb bus (the one before the
branched address).

This fixes this issue.

Patch by: Franck Jullien <franck.jullien@gmail.com>

git-svn-id: http://opencores.org/ocsvn/openrisc/openrisc/trunk/or1200@847 04c7f051-eaf3-4997-922d-f29d9d8c0f3f